Add nmap support for bulk requests
This assumes nmap is already installed and on the user's $PATH. Honestly pretty proud of getting this thing done, it was a lot of slogging through XML but we did it. The existing nmap libraries didn't play well with sudo-prompt, so a bit of rolling my own was needed. Unfortunately.